fix capitalization, un-title last paragraph
BitPay and GitHub are always spelled with camel case unless they are a part of
a URL or ID or some kind.

I also changed the last line so that it is not a title, which I just think
looks more professional if it is a sentence like normal.

I didn't notice any grammatical or spelling errors. Everything else looks good
to me.